% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_leap.R
\name{is_leap}
\alias{is_leap}
\title{Is this a leap year?}
\usage{
is_leap(num)
}
\arguments{
\item{num}{a number}
}
\value{
TRUE or FALSE on whether or not it is a leap year
}
\description{
Is this a leap year?
}
